get_user_group
